% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cohort_expectedCI_VE_irr.R
\name{get_cohort_expectedCI_VE_irr}
\alias{get_cohort_expectedCI_VE_irr}
\title{Expected lower and expected upper limit of the confidence intervals of variant and vaccine-specific efficacy based on incidence rate ratio.}
\usage{
get_cohort_expectedCI_VE_irr(
  anticipated_VE_for_each_brand_and_variant = matrix(data = c(0.1, 0.2, 0.3, 0.4, 0.5,
    0.6, 0.7, 0.8, 1), nrow = 3, ncol = 3, byrow = F, dimnames = list(paste0("brand",
    1:3), paste0("variant", 1:3))),
  brand_proportions_in_vaccinated = c(brand1 = 0.3, brand2 = 0.5, brand3 = 0.2),
  overall_vaccine_coverage = 0.3,
  proportion_variants_in_unvaccinated_cases = c(variant1 = 0.6, variant2 = 0.3,
    variant3 = 0.1),
  overall_attack_rate_in_unvaccinated = 0.1,
  study_period = 1,
  calculate_relative_VE = T,
  alpha = 0.05,
  confounder_adjustment_Rsquared = 0,
  prob_missing_data = 0.1,
  total_subjects = seq(1000, 10000, 25),
  nsims = 500
)
}
\arguments{
\item{anticipated_VE_for_each_brand_and_variant}{a matrix of vaccine efficacy of each vaccine (row)  against each variant (column). Each value must be a real number between 0 and 1.}

\item{brand_proportions_in_vaccinated}{a vector denoting the proportion in which vaccines are given in the vaccinated subjects of the study cohort. Each value of this vector must be a real number between 0 and 1 and the sum of the values of this vector must be equal to 1.}

\item{overall_vaccine_coverage}{the proportion of the study cohort that will be vaccinated. It should be a real number between 0 and 1.}

\item{proportion_variants_in_unvaccinated_cases}{a vector of the proportions in which each variant is expected to be present in the unvaccinated and infected subjects in the study cohort. Each value of this vector must be a real number between 0 and 1 and the sum of the values of this vector must be equal to 1.}

\item{overall_attack_rate_in_unvaccinated}{the proportion of the study cohort that is expected to infected over the study period. It should be a real number between 0 and 1.}

\item{study_period}{the study period (epochs) should be numeric value greater than 0.}

\item{calculate_relative_VE}{a logical indicating if calculations should also be done for relative vaccine efficacy (default \code{TRUE}).}

\item{alpha}{controls the width \verb{[100*alpha/2, 100*(1 - alpha/2)]\%} of the confidence interval. It is a numeric that must take a value between 0 and 1.}

\item{confounder_adjustment_Rsquared}{we use this parameter to adjust the calculations for potential confounders using the methodology proposed by Hsieh and Lavori (2000). It represents the amount of variance (R^2) explained in a regression model where vaccination status is the outcome and confounders of interest are predictors. It is a numeric that must take a value between 0 (no adjustment for confounders) and 1.}

\item{prob_missing_data}{to adjust the calculations for non-informative and random subject loss to follow-up/dropout. it should take a numeric value between 0 and 1.}

\item{total_subjects}{a vector of study cohort size for which calculations should be done.}

\item{nsims}{total number of Monte Carlo simulations conducted.}
}
\value{
A data frame consisting of the input parameters, absolute and relative VE combinations, and the expected lower and expected upper width of the confidence intervals for each absolute and relative VE.
}
\description{
The function \code{get_cohort_expectedCI_VE_irr} simulates confindence intervals for variant and vaccine-specific efficacy (VE) for a given sample size.
The efficacy is defined as \verb{VE = 1 - incidence rate ratio}, and the function returns expected lower and expected upper confidence interval limit for both absolute and relative VE.
}
\details{
#' In this function efficacy is defined as \verb{VE = 1 - incidence rate ratio}, where 'incidence rate ratio' is
the ratio of incidence rates of being a case of a particular variant/variant among the groups being compared.
When the groups being compared are a particular vaccine versus placebo then we call the VE
as the absolute VE of the vaccine. For \code{M} vaccines there are \code{M} absolute VE, one each for the \code{M} vaccines.
When the groups being compared are a particular vaccine versus another vaccine then we call the VE
as the relative VE of the vaccines, for a particular variant. For \code{M} vaccines and \code{I} variants there are \verb{I x 2 x utils::combn(M, 2)}
permutations of relative VE of two vaccines against the same variant.

We first transform the user inputs for \code{I} variants and \code{M} vaccines into a \verb{(I + 1) x (M + 1)} cross table of
cumulative-risks of being a case or a control over the study period. The overall sum of all cumulative-risks,
i.e., all cells, of this table is 1. The first row of our cumulative-risk table contain cumulative-risk of being a control.
The first column corresponds to subjects who are unvaccinated.
Thus, the cell \verb{\{1,1\}} contains the probability (cumulative-risk) that over the study period a subject will be a control and unvaccinated.
The remaining \code{Ì} rows correspond to subjects who are cases of a particular variant/variant of the pathogen,
and the remaining \code{M} columns correspond to subjects who are vaccinated with a particular vaccine.
The next step is to simulate the data. To speed up our computations we sample an \verb{(I + 1) x (M + 1)}
cross table of data from a multinomial distribution with probabilities taken from our cumulative-risk table.
The total subjects sampled in the cross table are are \code{total_subjects * (1 - prob_missing_data)}.

In the person-time based estimator of incidence rates the number of cases only constitute the numerator.
The denominator is person-time contributed by subjects vaccinated with a particular vaccine (or placebo).
Hence, in addition to the table of cumulative-risks, using the user input we also obtain a \verb{I x (M + 1)} table of incidence rates.
We use this table to calculate the person-time contribution of subjects over \code{study_period}.
In this table each cell contains the incidence rate of being a case of a particular variant given their vaccination status.
The naive method to calculate the overall person-time contribution of subjects vaccinated with a particular vaccine is to
simulate event-time of each subject using an exponential distribution (R function rexp). Subsequently, one can add them up over the subjects.
But this is computationally very slow. Hence we instead exploit the central limit theorem to directly sample the
'sum of the person-time of all subjects with a certain vaccination status'.
For this, assume that the \code{study_period=t} and the sum of the event rates
for \code{I} variants is \code{p = p1 + p2 + ... + pI} per unit time, where \verb{p1, p2, ..., pI} are the individual
event rates of the \code{I} variants. Then \verb{exp(-pt)\%} subjects will not be infected with the variant
and contribute \code{t} units time each. The remaining \verb{100 - exp(-pt)\%} subjects will
obtain the event and individually contribute an event time whose sum can be directly sampled from a
normal distribution (central limit theorem). This is because the sum of event times of \code{K}
eventful subjects follows a normal distribution with mean equal to
\verb{K x mean of a truncated exponential distribution in the interval} \verb{[0, t]}
and variance equal to the \verb{K x variance of a truncated exponential distribution in the interval} \verb{[0, t]}.

We then estimate the absolute and relative VE of each vaccine using the counts and person-times based on the sampled data.
For this purpose we also reduce the sample size by a factor \code{prob_missing_data * confounder_adjustment_Rsquared}.
The confidence intervals with widths \verb{[100*alpha/2, 100*(1 - alpha/2)]\%} are obtained using normal approximation
to the distribution of log of incidence rate ratio (Rothman KJ, 2012).
To adjust for confounders, the standard-error used in the confidence interval is rescaled to \code{SE/(1 - confounder_adjustment_Rsquared)} (Hsieh and Lavori, 2000)
We repeat this procedure \code{nsims} times, and in each such simulation we obtain \code{nsims} confidence intervals.

To conduct simulations faster all the calculations are done without using for loops.
Instead we use a three-dimensional R arrays, with one-dimension for \code{nsims},
another for the sample size vector \code{total_subjects},
and another for a vector containing the flattened cross table of simulated data on cases and controls.
}
\references{
\enumerate{
\item Hsieh, F. Y., & Lavori, P. W. (2000). Sample-size calculations for the Cox proportional hazards regression model with nonbinary covariates. Controlled clinical trials, 21(6), 552-560.
\item Rothman KJ (2012) Epidemiology: An Introduction. 2nd Ed., Oxford University Press, Oxford.
}
}
