% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_leaky_ve_estimands.R
\name{get_leaky_ve_estimands}
\alias{get_leaky_ve_estimands}
\title{Time-dependency of variant and vaccine-specific efficacy (VE) estimands based on incidence rate ratio, cumulative-incidence ratio, and odds ratio}
\usage{
get_leaky_ve_estimands(
  anticipated_VE_for_each_brand_and_variant = matrix(data = c(0.1, 0.3, 0.4, 0.6, 0.7,
    0.9), nrow = 2, ncol = 3, byrow = F, dimnames = list(paste0("brand", 1:2),
    paste0("variant", 1:3))),
  incidence_rate_unvaccinated = c(0.05, 0.1, 0.2),
  study_period = 1
)
}
\arguments{
\item{anticipated_VE_for_each_brand_and_variant}{a matrix of vaccine efficacy of each vaccine (row) against each variant (column). Each value must be a real number between 0 and 1.}

\item{incidence_rate_unvaccinated}{a vector denoting the incidence rates of each variant in the unvaccinated subjects.}

\item{study_period}{the study period (epochs) should be numeric value greater than 0.}
}
\value{
A list with three elements, namely \code{absolute_ve}, \code{relative_ve_across_variant_given_vaccine},
and \code{relative_ve_across_vaccines_given_variant}.
The list element \code{absolute_ve} pertains to the absolute efficacy of the vaccines.
It contains four elements, namely \code{absolute_ve_true}, \code{absolute_ve_irr},
\code{absolute_ve_cir}, and \code{absolute_ve_or}. Here \code{absolute_ve_true} is same as the input parameter
\code{anticipated_VE_for_each_brand_and_variant}, and indicates the true biological efficacy of the vaccine.
Whereas, the remaining three are the VE estimands based on incidence rate ratio, cumulative-incidence ratio,
and odds ratio, respectively.

The list element \code{relative_ve_across_variant_given_vaccine} pertains to the relative VE of a vaccine against two different
variants/variants. It is a matrix with 7 rows and multiple columns. Each column being one of the comparison sets of variants and vaccines.
Rows 1 and 2 indicate the two variants of interest for comparison and row 3 has the vaccine.
The remaining rows show the true relative VE, and the relative VE based on incidence rate ratio, cumulative-incidence ratio,
and odds ratio, respectively.

The list element \code{relative_ve_across_vaccines_given_variant} pertains to the relative VE of two vaccines against the
same variants/variants. It is a matrix with 7 rows and multiple columns.
Each column being one of the comparison sets of variants and vaccines.
Rows 1 and 2 indicate the two vaccines of interest for comparison and row 3 has the variant.
The remaining rows show the true relative VE, and the relative VE based on incidence rate ratio, cumulative-incidence ratio,
and odds ratio, respectively.
}
\description{
The function \code{get_leaky_ve_estimands} provides the VE estimands based on incidence rate ratio, cumulative-incidence ratio, and odds ratio
for a leaky vaccine given the actual efficacy of the vaccine. The function returns both absolute and relative VE estimands.
}
\details{
A leaky vaccine is characterized by its ability to reduce the force of infection \code{lambda} in unvaccinated subjects to factor
\code{theta * lambda} in vaccinated subjects, where \verb{0 <= 1 - theta <=1 }is the true VE of the leaky vaccine. This true VE
is to be entered in the matrix anticipated_VE_for_each_brand_and_variant.

Smith et al. (1984) have shown that when a vaccine has a leaky action mechanism (Halloran et al., 2010, page 132),
then VE calculated as \verb{VE = 1 - cumulative-incidence ratio} depends on the length of the study period \code{t}.
Specifically, when \code{t -> Infinity} then \verb{1 - cumulative-incidence ratio -> 0} and thus it does not reflect the
biological effect of the vaccine for a given exposure to the pathogen. The aim of this function is to
extend upon the findings of Smith et al. (1984) for multiple variant/variants and vaccines. And to help
practitioners understand why certains VE measures may not reflect the biological VE.
More details can be found in our related paper.
}
\examples{
As an example we recommend running the function without passing any parameter to it.
The default scenario is for three vaccines and three pathogen variants.

}
\references{
\enumerate{
\item Halloran, M. E., Longini, I. M., Struchiner, C. J., and Longini, I. M. (2010).Design and analysis of vaccine studies, volume 18. Springer.
\item Smith, P., Rodrigues, L., and Fine, P. (1984). Assessment of the protective efficacy of vaccines against common diseases using case-control and cohort studies.International journal of epidemiology, 13(1):87–93.
}
}
