% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cohort_expectedCI_VE_cir.R
\name{get_cohort_expectedCI_VE_cir}
\alias{get_cohort_expectedCI_VE_cir}
\title{Get expected lower and expected upper limit of strain and vaccine-specific efficacy}
\usage{
get_cohort_expectedCI_VE_cir(
  anticipated_VE_for_each_brand_and_strain = matrix(data = c(0.1, 0.2, 0.3, 0.4, 0.5,
    0.6, 0.7, 0.8, 1), nrow = 3, ncol = 3, byrow = T, dimnames = list(paste0("brand",
    1:3), paste0("strain", 1:3))),
  brand_proportions_in_vaccinated = c(brand1 = 0.3, brand2 = 0.5, brand3 = 0.2),
  overall_vaccine_coverage = 0.3,
  proportion_strains_in_unvaccinated_cases = c(strain1 = 0.6, strain2 = 0.3, strain3 =
    0.1),
  overall_attack_rate_in_unvaccinated = 0.1,
  calculate_relative_VE = T,
  alpha = 0.05,
  confounder_adjustment_Rsquared = 0,
  prob_missing_data = 0.1,
  total_subjects = seq(1000, 10000, 25),
  nsims = 500
)
}
\arguments{
\item{anticipated_VE_for_each_brand_and_strain}{a matrix of vaccine efficacy of each vaccine (column) against each strain (row). Each value must be a real number between 0 and 1.}

\item{brand_proportions_in_vaccinated}{a vector denoting the proportion in which vaccines are given in the vaccinated subjects of the study cohort. Each value of this vector must be a real number between 0 and 1 and the sum of the values of this vector must be equal to 1.}

\item{overall_vaccine_coverage}{the proportion of the study cohort that will be vaccinated. It should be a real number between 0 and 1.}

\item{proportion_strains_in_unvaccinated_cases}{a vector of the proportions in which each strain is expected to be present in the unvaccinated and infected subjects in the study cohort. Each value of this vector must be a real number between 0 and 1 and the sum of the values of this vector must be equal to 1.}

\item{overall_attack_rate_in_unvaccinated}{the proportion of the study cohort that is expected to infected over the study period. It should be a real number between 0 and 1.}

\item{calculate_relative_VE}{a logical indicating if calculations should also be done for relative vaccine efficacy (default \code{TRUE}).}

\item{alpha}{It controls the width \verb{[100*alpha/2, 100*(1 - alpha/2)]\%} of the confidence interval. It is a numeric that must take a value between 0 and 1. .}

\item{confounder_adjustment_Rsquared}{we use this parameter to adjust the calculations for potential confounders using the methodology proposed by \insertCite{hsieh2000sample}. It represents the amount of variance (R^2) explained in a regression model where vaccination status is the outcome and confounders of interest are predictors. It is a numeric that must take a value between 0 (no adjustment for confonders) and 1.}

\item{prob_missing_data}{to adjust the calculations for non-informative and random subject loss to follow-up/dropout. it should take a numeric value between 0 and 1.}

\item{total_subjects}{a vector of study cohort size for which calculations should be done.}

\item{nsims}{total number of Monte Carlo simulations conducted.}
}
\description{
The function \code{get_cohort_expectedCI_VE_cir} returns the sum of all the values present in its arguments.
}
\references{
\preformatted{\\insertAllCited\{\}
}
}
